requirements.txt
# Framework web
streamlit==1.30.0
aiohttp==3.9.1

# Bot Framework
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botbuilder-integration-aiohttp==4.15.0

# Azure AI SDK
azure-ai-ml==1.12.0
azure-identity==1.15.0
azure-ai-inference==1.0.0b1

# LangChain
langchain==0.1.4
langchain-openai==0.0.5
openai==1.10.0

# Azure OpenAI (para conexión directa si es necesario)
azure-openai==1.3.0

# Content Safety
azure-ai-contentsafety==1.0.0

# Utilidades
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.5.3
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0

# Adaptive Cards
adaptivecards==0.1.0
